+++ src/ansys/fluent/core/exceptions.py
@@ -22,7 +22,6 @@
 
 """Custom common higher level exceptions."""
 from typing import Any, Iterable
-import warnings
 
 from ansys.fluent.core.solver.error_message import allowed_name_error_message
 
@@ -51,15 +50,3 @@ class InvalidArgument(ValueError):
     """Raised when an argument value is inappropriate."""
 
     pass
-
-
-def warning_for_fluent_dev_version(version):
-    """Provides warning if Fluent develop branch is used."""
-    from ansys.fluent.core import FLUENT_RELEASE_VERSION, FluentVersion
-
-    if FluentVersion(version) > FluentVersion(FLUENT_RELEASE_VERSION):
-        warnings.warn(
-            "⚠️ Warning: You are using PyFluent with an unreleased or development version of Fluent.\n"
-            "Compatibility is not guaranteed, and unexpected behavior may occur. Please use a released "
-            "version of Fluent that is officially supported by this version of PyFluent."
-        )
